+++ core/ui_util/src/main/java/com/napzak/market/ui_util/NavGraphBuilderExt.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.util.android
+package com.napzak.market.ui_util
 
 import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.core.tween
