+++ feature/mypage/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.common)
     implementation(projects.core.util)
+    implementation(projects.core.uiUtil)
     implementation(projects.domain.store)
     implementation(projects.feature.detail)
